+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes of this package are documented in this file using the [Keep 
 ### Changed
 
 * [wiz] `:init` and `:config` now accept multiple lines without needing to be enclosed in `lambda`.
+* [wiz] Make delayed require when use `:package` keyword
 
 ## [0.0.4]